python_workers/s3_worker/requirements.txt
redis
boto3
python-dotenv
PyYAML
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-exporter-otlp
opentelemetry-instrumentation-botocore
jsonschema
